Switch to scantxoutset for 2of3 recovery
If, while performing the 2of3 recovery, the user is
running core with version 0.17.00 and does
not want to scan the mempool (by specifying
--ignore-mempool), then he can use scantxoutset
(instead of importmulti + listunspent), which is
faster.

To scan the mempool is necessary to use importmulti +
listunspent or wait for scanmempool.

Choosing "addr(<address>)" as descriptor was done
considering that the csv script could not delegate
the HD derivation to core.