Add support for bip39 passphrase
Only applies if passing mnemonic words (as opposed to explicit seed).

Note: mnemonic is also used for deriving potential gait path for legacy
wallets - passphrase is not used in this derivation.  At the time we
did not support bip39 passphrase, so any wallets that have a gait path
derived from mnemonic would not have a bip39 passphrase.